feat: toggle stack in StackGRU
Description: Allows for the stack to be turned off (so use just as "vanilla" GRU). This (at the moment) is mainly intended to be a test feature.

This commit also adds a test case for the GRU w/o Stack. It is intended to allow us to locate those cases where the Stack makes the model crash in a weird and silent mode.