Speed improvement (#12)
* wip: pushing stashed workstate

* test: add speed unittest for StackGRU/VAE

* test: add speed unittest for StackGRU/VAE

* wip: add packed sequences to data preparation

Description: Using a packed sequence may reduce
the number of operations that we will need to do
when running the model.

* wip: gru forward with packed sequences

Related to #11

Description: (WIP) Allows to run all the steps of the unidirectional
encoder GRU with packed sequences. It will likely break things as it
is right now with any other configuration. Stack is not yet working
so it also needs to be disabled for this commit to work.

* wip: allow packed bidirectional

NOTE: The code in this commit will likely not run.

* feat: packed sequences on GRU

NOTES: Needs tests & refactoring. Stack not working
yet.

* feat: add stacks to packed mode

* chore: reverse input using tensors

* wip: packed input pipeline

* chore: refactor forward functions

* feat: add prepare functions to training fns

* fix: add missing return and test param

* wip: misc. fix

* test: Packed speed test

This is a happy moment :)

* fix: missing assert

* chore: reduce number of params in speed test

* refactor: stackRNN adaptations

* refactor: move out output_layer from stackgru

* refactor: remove init_stack, init_hidden and batch processign from stackgru forward call

* feat: add logging for reconstructetd mol

* wip: new training script

* test/fix: shrink down #test

* wip: remove addititon of start token and cleanup data preparation

* chore

Co-authored-by: Nil Adell Mill <nil.adell01@gmail.com>
Co-authored-by: Nil Adell Mill <nila@ethz.ch>